feat: continued work with endpoints
had and fixed, for now, trouble with Lazy fetching (Section.arifactsList now has EAGER fetching) - WIP
added a few endpoints, will continue work
