+++ src/main/java/ro/ubb/abc2024/arheo/controller/section/SectionController.java
@@ -1,0 +1,64 @@
+package ro.ubb.abc2024.arheo.controller.section;
+
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.*;
+import ro.ubb.abc2024.arheo.service.SectionService;
+import ro.ubb.abc2024.arheo.utils.converter.SectionDtoConverter;
+import ro.ubb.abc2024.arheo.utils.dto.SectionDto;
+import ro.ubb.abc2024.user.Role;
+import ro.ubb.abc2024.utils.dto.Result;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("${api.endpoint.arheo}/sections")
+public class SectionController {
+    private final SectionService sectionService;
+    private final SectionDtoConverter sectionDtoConverter;
+
+    @GetMapping
+    //@PreAuthorize("hasAnyAuthority('ARH')")
+    // get all sections
+    public Result<List<SectionDto>> getSections() {
+        return new Result<>(true, HttpStatus.OK.value(), "Retrieved all sections", sectionService.getSections().stream()
+                .map(sectionDtoConverter::createFromEntity)
+                .collect(Collectors.toList()));
+    }
+
+    // get by id
+    @GetMapping("/{id}")
+    //@PreAuthorize("hasAnyAuthority('ARH')")
+    public Result<SectionDto> getSection(@PathVariable long id) {
+        return new Result<>(true, HttpStatus.OK.value(), "Retrieved section", sectionDtoConverter.createFromEntity(sectionService.getSection(id)));
+    }
+
+    @PostMapping
+    //@PreAuthorize("hasAnyAuthority('ARH')")
+    // add a new section, add it to the list of sections
+    public Result<SectionDto> addSection(@RequestBody SectionDto sectionDto) {
+        var section = sectionService.addSection(sectionDtoConverter.createFromDto(sectionDto));
+        //return sectionDtoConverter.createFromEntity(section);
+        return new Result<>(true, HttpStatus.CREATED.value(), "Added section", sectionDtoConverter.createFromEntity(section));
+    }
+
+    @DeleteMapping
+    //@PreAuthorize("hasAnyAuthority('ARH')")
+    public Result<SectionDto> removeSection(@RequestBody long sectionId){
+        // given a section id, remove the section from the list of sections
+        sectionService.deleteSection(sectionId);
+        return new Result<>(true, HttpStatus.OK.value(), "Deleted section", sectionDtoConverter.createFromEntity(sectionService.getSection(sectionId)));
+    }
+
+    //@PutMapping
+    //@PreAuthorize("hasAnyAuthority('ARH')")
+    // edit/update
+}
+
+
+
