+++ src/main/java/ro/ubb/abc2024/arheo/domain/section/Section.java
@@ -100,7 +100,7 @@ public class Section {
 
     // list of artifacts, nullable; reference to the artifacts
     // list of artifacts found in the section
-    @OneToMany(targetEntity = Artifact.class, mappedBy = "section", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
+    @OneToMany(targetEntity = Artifact.class, mappedBy = "section", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.EAGER)
     // join column correctly; artifacts are expected to have a section_id
     private List<Artifact> artifactsList;
