+++ src/main/java/ro/ubb/abc2024/arheo/service/SectionService.java
@@ -1,5 +1,6 @@
 package ro.ubb.abc2024.arheo.service;
 
+import jakarta.transaction.Transactional;
 import ro.ubb.abc2024.arheo.domain.artifact.Artifact;
 import ro.ubb.abc2024.arheo.domain.section.Section;
 import java.util.List;
@@ -16,6 +17,7 @@ public interface SectionService {
     List<Section> getSectionsSouthOf(double latitude);
     List<Section> getSectionsWestOf(double longitude);
     List<Section> getSectionsEastOf(double longitude);
+    @Transactional
     List<Section> getSections();
     List<Section> getSectionsBySite(long siteId);
     List<Section> getSectionsByStatusIsAndSiteId(String status, long siteId);
